+++ src/main/java/moonshine/core/events/clientTick.java
@@ -1,0 +1,10 @@
+package moonshine.core.events;
+
+import net.minecraftforge.fml.common.Mod;
+import moonshine.core.moonshine;
+
+@Mod.EventBusSubscriber(modid = moonshine.ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
+public class clientTick
+{
+
+}
