+++ src/main/java/moonshine/util/Process.java
@@ -1,0 +1,6 @@
+package moonshine.util;
+
+@FunctionalInterface
+public interface Process<T> {
+	T process(T t);
+}
