--- src/main/java/tac_extra/core/events/clientTick.java
@@ -1,10 +1,0 @@
-package tac_extra.core.events;
-
-import net.minecraftforge.fml.common.Mod;
-import tac_extra.core.tac_extra;
-
-@Mod.EventBusSubscriber(modid = tac_extra.ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
-public class clientTick
-{
-
-}
