--- src/main/java/tac_extra/util/Process.java
@@ -1,6 +1,0 @@
-package tac_extra.util;
-
-@FunctionalInterface
-public interface Process<T> {
-	T process(T t);
-}
