+++ src/main/resources/assets/moonshine/lang/en_au.json
@@ -1,0 +1,84 @@
+{
+  "item.moonshine.m1911": "M1911",
+  "item.moonshine.m1894": "M1894",
+  "item.moonshine.m1851": "M1851",
+  "item.moonshine.round45": ".45 ACP Bullet",
+  "item.moonshine.magnumround": ".357 Magnum Bullet",
+  "item.moonshine.win_30-30": ".30-30 Winchester",
+  "item.moonshine.box45": "Box of .45 ACP Bullets",
+  "item.moonshine.magnumbox": "Box of .357 Magnum Bullets",
+  "item.moonshine.m1928": "M1928",
+  "item.moonshine.mosin": "Mosin Nagant",
+  "item.moonshine.ak47": "AK-47",
+  "item.moonshine.bullet_308": ".308 Bullet",
+  "item.moonshine.box_308-block": "Box of .308 Bullets",
+  "item.moonshine.m60": "M60",
+  "item.moonshine.m1917": "M1917",
+  "item.moonshine.glock_17": "Glock 17",
+  "item.moonshine.dp28": "DP-28",
+  "item.moonshine.m16a1": "M16 A1",
+  "item.moonshine.mk18": "MK 18",
+  "item.moonshine.sti2011": "STI 2011",
+  "item.moonshine.ak74": "AK-74",
+  "item.moonshine.m92fs": "M92FS",
+  "item.moonshine.nato_556_bullet": "5.56 Bullet",
+  "item.moonshine.nato_556_box_block": "Box of 5.56 Bullets",
+  "item.moonshine.9mm_box_block": "Box of 9mm Bullets (x24)",
+  "item.moonshine.9mm_round": "9mm Bullet",
+  "item.moonshine.ar_15_p": "AR-15 P",
+  "item.moonshine.ar_15_hellmouth": ".308 Hellmouth",
+  "item.moonshine.micro_uzi": "Micro Uzi",
+  "item.moonshine.vector45": "Vector 45",
+  "item.moonshine.walther_ppk": "Walther PPK",
+  "item.moonshine.mosberg590": "Mosberg 590",
+  "item.moonshine.db_long": "Double Barrel",
+  "item.moonshine.db_short": "Sawed off Double Barrel",
+  "item.moonshine.10_gauge_round": "10 Gauge Round",
+  "item.moonshine.m4": "M4A1",
+  "item.moonshine.m24": "M24",
+  "item.moonshine.m1911_nether": " Nether M1911",
+
+  "item.moonshine.timeless_workbench": "Rustic Workbench",
+
+  "gui.moonshine.timeless_workbench.assemble": "Assemble",
+  "gui.moonshine.timeless_workbench.show_remaining": "Show Remaining",
+  "gui.moonshine.timeless_workbench.tab.weapons": "Weapons",
+  "gui.moonshine.timeless_workbench.tab.attachments": "Attachments",
+  "gui.moonshine.timeless_workbench.tab.ammo": "Ammo",
+  "gui.moonshine.timeless_workbench.tab.misc": "Ammo Boxes",
+  "container.moonshine.timeless_workbench": "",
+
+  "block.moonshine.box45": "Box of .45 ACP Bullets (x16)",
+  "block.moonshine.magnumbox": "Box of .357 Magnum Bullets (x12)",
+  "block.moonshine.win30-30box": "Box of .30-30 Winchester Bullets (x10)",
+  "block.moonshine.box_308-block": "Box of .308 Bullets (x50)",
+  "block.moonshine.timeless_workbench": "Rustic Workbench",
+  "block.moonshine.nato_556_box_block": "Box of 5.56 Bullets (x50)",
+  "block.moonshine.9mm_box_block": "Box of 9mm Bullets (x24)",
+  "block.moonshine.10_gauge_box_block": "Box of 10g Bullets (x12)",
+
+  "itemGroup.moonshine": "Timeless and Classics!",
+  "itemGroup.timeless_and_modern": "Timeless and Modern!",
+
+  "subtitle.moonshine.fire_1911fire": "45 Pistol Fire",
+  "subtitle.moonshine.fire_1911fires": "45 Pistol Fire Silenced",
+  "subtitle.moonshine.fire_mosin-nagantfire": "Mosin Fire",
+  "subtitle.moonshine.fire_30-30fire": "30-30 Fire",
+  "subtitle.moonshine.fire_ak74fire": "AK-74 Fire",
+  "subtitle.moonshine.fire_ak74fires": "AK-74 Fire Silenced",
+  "subtitle.moonshine.fire_ak47fire": "AK-47 Fire",
+  "subtitle.moonshine.fire_ak47fires": "AK-47 Fire Silenced",
+  "subtitle.moonshine.fire_dp28fire": "DP-28 Fire",
+  "subtitle.moonshine.fire_glockfire": "Glock Fire",
+  "subtitle.moonshine.fire_glockfires": "Glock Fire Silenced",
+  "subtitle.moonshine.fire_m16fire": "AR Fire",
+  "subtitle.moonshine.fire_m16fires": "AR Fire Silenced",
+  "subtitle.moonshine.fire_m60fire": "M60 Fire",
+  "subtitle.moonshine.fire_m92fsfires": "M9 Fire Silenced",
+  "subtitle.moonshine.fire_m92fsfire": "M9 Fire",
+  "subtitle.moonshine.fire_m1851fire": "M1851 Fire Silenced",
+  "subtitle.moonshine.fire_tompsonfire": "Thompson Fire",
+  "subtitle.moonshine.fire_dbfire": "Shotgun Fire",
+
+  "info.moonshine.auto": "This weapon is automatic!"
+}
