+++ src/main/resources/assets/moonshine/sounds/item/30-30_fire.ogg
