+++ src/main/resources/assets/moonshine/textures/gui/workbench.png
