+++ src/main/resources/assets/moonshine/textures/items/m1894_body.png
