+++ src/main/resources/assets/moonshine/textures/items/win_30-30_bullet.png
