--- src/main/resources/assets/tac_extra/lang/en_ca.json
@@ -1,84 +1,0 @@
-{
-  "item.tac_extra.m1911": "M1911",
-  "item.tac_extra.m1894": "M1894",
-  "item.tac_extra.m1851": "M1851",
-  "item.tac_extra.round45": ".45 ACP Bullet",
-  "item.tac_extra.magnumround": ".357 Magnum Bullet",
-  "item.tac_extra.win_30-30": ".30-30 Winchester",
-  "item.tac_extra.box45": "Box of .45 ACP Bullets",
-  "item.tac_extra.magnumbox": "Box of .357 Magnum Bullets",
-  "item.tac_extra.m1928": "M1928",
-  "item.tac_extra.mosin": "Mosin Nagant",
-  "item.tac_extra.ak47": "AK-47",
-  "item.tac_extra.bullet_308": ".308 Bullet",
-  "item.tac_extra.box_308-block": "Box of .308 Bullets",
-  "item.tac_extra.m60": "M60",
-  "item.tac_extra.m1917": "M1917",
-  "item.tac_extra.glock_17": "Glock 17",
-  "item.tac_extra.dp28": "DP-28",
-  "item.tac_extra.m16a1": "M16 A1",
-  "item.tac_extra.mk18": "MK 18",
-  "item.tac_extra.sti2011": "STI 2011",
-  "item.tac_extra.ak74": "AK-74",
-  "item.tac_extra.m92fs": "M92FS",
-  "item.tac_extra.nato_556_bullet": "5.56 Bullet",
-  "item.tac_extra.nato_556_box_block": "Box of 5.56 Bullets",
-  "item.tac_extra.9mm_box_block": "Box of 9mm Bullets (x24)",
-  "item.tac_extra.9mm_round": "9mm Bullet",
-  "item.tac_extra.ar_15_p": "AR-15 P",
-  "item.tac_extra.ar_15_hellmouth": ".308 Hellmouth",
-  "item.tac_extra.micro_uzi": "Micro Uzi",
-  "item.tac_extra.vector45": "Vector 45",
-  "item.tac_extra.walther_ppk": "Walther PPK",
-  "item.tac_extra.mosberg590": "Mosberg 590",
-  "item.tac_extra.db_long": "Double Barrel",
-  "item.tac_extra.db_short": "Sawed off Double Barrel",
-  "item.tac_extra.10_gauge_round": "10 Gauge Round",
-  "item.tac_extra.m4": "M4A1",
-  "item.tac_extra.m24": "M24",
-  "item.tac_extra.m1911_nether": " Nether M1911",
-
-  "item.tac_extra.timeless_workbench": "Rustic Workbench",
-
-  "gui.tac_extra.timeless_workbench.assemble": "Assemble",
-  "gui.tac_extra.timeless_workbench.show_remaining": "Show Remaining",
-  "gui.tac_extra.timeless_workbench.tab.weapons": "Weapons",
-  "gui.tac_extra.timeless_workbench.tab.attachments": "Attachments",
-  "gui.tac_extra.timeless_workbench.tab.ammo": "Ammo",
-  "gui.tac_extra.timeless_workbench.tab.misc": "Ammo Boxes",
-  "container.tac_extra.timeless_workbench": "",
-
-  "block.tac_extra.box45": "Box of .45 ACP Bullets (x16)",
-  "block.tac_extra.magnumbox": "Box of .357 Magnum Bullets (x12)",
-  "block.tac_extra.win30-30box": "Box of .30-30 Winchester Bullets (x10)",
-  "block.tac_extra.box_308-block": "Box of .308 Bullets (x50)",
-  "block.tac_extra.timeless_workbench": "Rustic Workbench",
-  "block.tac_extra.nato_556_box_block": "Box of 5.56 Bullets (x50)",
-  "block.tac_extra.9mm_box_block": "Box of 9mm Bullets (x24)",
-  "block.tac_extra.10_gauge_box_block": "Box of 10g Bullets (x12)",
-
-  "itemGroup.tac_extra": "Timeless and Classics!",
-  "itemGroup.timeless_and_modern": "Timeless and Modern!",
-
-  "subtitle.tac_extra.fire_1911fire": "45 Pistol Fire",
-  "subtitle.tac_extra.fire_1911fires": "45 Pistol Fire Silenced",
-  "subtitle.tac_extra.fire_mosin-nagantfire": "Mosin Fire",
-  "subtitle.tac_extra.fire_30-30fire": "30-30 Fire",
-  "subtitle.tac_extra.fire_ak74fire": "AK-74 Fire",
-  "subtitle.tac_extra.fire_ak74fires": "AK-74 Fire Silenced",
-  "subtitle.tac_extra.fire_ak47fire": "AK-47 Fire",
-  "subtitle.tac_extra.fire_ak47fires": "AK-47 Fire Silenced",
-  "subtitle.tac_extra.fire_dp28fire": "DP-28 Fire",
-  "subtitle.tac_extra.fire_glockfire": "Glock Fire",
-  "subtitle.tac_extra.fire_glockfires": "Glock Fire Silenced",
-  "subtitle.tac_extra.fire_m16fire": "AR Fire",
-  "subtitle.tac_extra.fire_m16fires": "AR Fire Silenced",
-  "subtitle.tac_extra.fire_m60fire": "M60 Fire",
-  "subtitle.tac_extra.fire_m92fsfires": "M9 Fire Silenced",
-  "subtitle.tac_extra.fire_m92fsfire": "M9 Fire",
-  "subtitle.tac_extra.fire_m1851fire": "M1851 Fire Silenced",
-  "subtitle.tac_extra.fire_tompsonfire": "Thompson Fire",
-  "subtitle.tac_extra.fire_dbfire": "Shotgun Fire",
-
-  "info.tac_extra.auto": "This weapon is automatic!"
-}
