--- src/main/resources/assets/tac_extra/models/item/9mm_box_block.json
@@ -1,53 +1,0 @@
-{
-	"credit": "Made with Blockbench",
-	"textures": {
-		"2": "tac_extra:block/box9",
-		"particle": "tac_extra:block/box9"
-	},
-	"elements": [
-		{
-			"from": [4, 0, 5],
-			"to": [12, 5, 10],
-			"faces": {
-				"north": {"uv": [0, 0, 8, 5], "texture": "#2"},
-				"east": {"uv": [0, 10, 5, 15], "texture": "#2"},
-				"south": {"uv": [0, 5, 8, 10], "texture": "#2"},
-				"west": {"uv": [0, 10, 5, 15], "texture": "#2"},
-				"up": {"uv": [8, 0, 16, 5], "texture": "#2"},
-				"down": {"uv": [8, 5, 16, 10], "texture": "#2"}
-			}
-		}
-	],
-	"display": {
-		"thirdperson_righthand": {
-			"rotation": [0, 180, 0],
-			"translation": [0, 3.75, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"thirdperson_lefthand": {
-			"rotation": [0, 180, 0],
-			"translation": [0, 3.75, 1.5],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"firstperson_righthand": {
-			"rotation": [0, -155, 0],
-			"translation": [2.5, 7, -2]
-		},
-		"firstperson_lefthand": {
-			"rotation": [0, -155, 0],
-			"translation": [2.5, 7, -2]
-		},
-		"ground": {
-			"translation": [0, 1, 0],
-			"scale": [0.5, 0.5, 0.5]
-		},
-		"gui": {
-			"rotation": [30, -145, 0],
-			"translation": [0, 5.75, 0],
-			"scale": [1.2, 1.2, 1.2]
-		},
-		"fixed": {
-			"translation": [0, 5.5, -2.5]
-		}
-	}
-}
