--- src/main/resources/assets/tac_extra/models/item/9mm_round.json
@@ -1,125 +1,0 @@
-{
-	"credit": "Made with Blockbench",
-	"texture_size": [32, 32],
-	"textures": {
-		"2": "tac_extra:items/9mm",
-		"particle": "models:cartridge/9mm"
-	},
-	"elements": [
-		{
-			"from": [7.78516, 0.13673, 7.78516],
-			"to": [8.21485, 0.9375, 8.21485],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [0, 5, 4, 12], "texture": "#2"},
-				"east": {"uv": [0, 5, 4, 12], "texture": "#2"},
-				"south": {"uv": [0, 5, 4, 12], "texture": "#2"},
-				"west": {"uv": [0, 5, 4, 12], "texture": "#2"},
-				"up": {"uv": [0, 0, 4, 4], "texture": "#2"},
-				"down": {"uv": [4, 11, 5, 12], "texture": "#2"}
-			}
-		},
-		{
-			"from": [7.78516, 0, 7.78516],
-			"to": [8.21485, 0.0586, 8.21485],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [4, 11, 9, 12], "texture": "#2"},
-				"east": {"uv": [4, 11, 9, 12], "texture": "#2"},
-				"south": {"uv": [4, 11, 9, 12], "texture": "#2"},
-				"west": {"uv": [4, 11, 9, 12], "texture": "#2"},
-				"up": {"uv": [0, 0, 0.5, 0.5], "texture": "#2"},
-				"down": {"uv": [4, 5, 10, 11], "texture": "#2"}
-			}
-		},
-		{
-			"from": [7.84375, 0.0586, 7.84375],
-			"to": [8.15625, 0.13673, 8.15625],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [4, 11, 8, 12], "texture": "#2"},
-				"east": {"uv": [4, 11, 8, 12], "texture": "#2"},
-				"south": {"uv": [4, 11, 8, 12], "texture": "#2"},
-				"west": {"uv": [4, 11, 8, 12], "texture": "#2"},
-				"up": {"uv": [0, 0, 0.375, 0.375], "texture": "#2"},
-				"down": {"uv": [0, 0, 0.375, 0.375], "texture": "#2"}
-			}
-		},
-		{
-			"from": [7.83791, 1.13281, 7.83791],
-			"to": [8.16212, 1.28906, 8.16212],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"east": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"south": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"west": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"up": {"uv": [0, 12, 4, 16], "texture": "#2"},
-				"down": {"uv": [0, 0, 0.3125, 0.3125], "texture": "#2"}
-			}
-		},
-		{
-			"from": [7.81446, 0.91798, 7.81446],
-			"to": [8.18555, 1.13281, 8.18555],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [4, 2, 10, 5], "texture": "#2"},
-				"east": {"uv": [4, 2, 10, 5], "texture": "#2"},
-				"south": {"uv": [4, 2, 10, 5], "texture": "#2"},
-				"west": {"uv": [4, 2, 10, 5], "texture": "#2"},
-				"up": {"uv": [10, 0, 16, 6], "texture": "#2"},
-				"down": {"uv": [4, 2, 10, 5], "texture": "#2"}
-			}
-		},
-		{
-			"from": [7.89259, 1.25, 7.89259],
-			"to": [8.10743, 1.38673, 8.10743],
-			"rotation": {"angle": 0, "axis": "y", "origin": [5, 0.70313, 5]},
-			"faces": {
-				"north": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"east": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"south": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"west": {"uv": [5, 0, 9, 2], "texture": "#2"},
-				"up": {"uv": [0, 12, 4, 16], "texture": "#2"},
-				"down": {"uv": [0, 0, 0.3125, 0.3125], "texture": "#2"}
-			}
-		}
-	],
-	"display": {
-		"thirdperson_righthand": {
-			"translation": [-0.5, 7.75, 0.25]
-		},
-		"thirdperson_lefthand": {
-			"translation": [-0.5, 7.75, 0.25]
-		},
-		"firstperson_righthand": {
-			"rotation": [-30, 0, 0],
-			"translation": [3.5, 7.5, -2.5]
-		},
-		"firstperson_lefthand": {
-			"rotation": [-30, 0, 0],
-			"translation": [3.5, 7.5, -2.5]
-		},
-		"ground": {
-			"translation": [0, 15, 0],
-			"scale": [2, 2, 2]
-		},
-		"gui": {
-			"rotation": [30, -145, 0],
-			"translation": [0, 24, 0],
-			"scale": [4, 4, 4]
-		},
-		"fixed": {
-			"rotation": [0, 90, 0],
-			"translation": [0, 14, -0.5],
-			"scale": [2, 2, 2]
-		}
-	},
-	"groups": [
-		{
-			"name": "9mm",
-			"origin": [0, 0, 0],
-			"children": [0, 1, 2, 3, 4, 5]
-		}
-	]
-}
