--- src/main/resources/assets/tac_extra/sounds.json
@@ -1,14 +1,0 @@
-{
-  "item.fire_30-30": {
-    "subtitle": "subtitle.tac_extra.fire_30-30fire",
-    "sounds": [
-      { "name": "tac_extra:item/30-30_fire" }
-    ]
-  },
-  "item.fire_30-30s": {
-    "subtitle": "subtitle.tac_extra.fire_30-30fires",
-    "sounds": [
-      { "name": "tac_extra:item/30-30_fires" }
-    ]
-  }
-}
