+++ src/main/resources/data/moonshine/guns/raiphala_skh_3.json
@@ -1,0 +1,65 @@
+{
+  "general": {
+    "boltAction": true,
+    "rate": 32,
+    "rateSelector": [
+      0,
+      1
+    ],
+    "gripType": "tac:two_handed_mosin",
+    "recoilAngle": 2.075,
+    "recoilKick": 1.95,
+    "horizontalRecoilAngle": 0.125,
+    "recoilAdsReduction": 0.40,
+    "weaponRecoilDuration": 0.85,
+    "recoilDuration": 0.125,
+    "cameraRecoilModifier": 2.0,
+    "alwaysSpread": true,
+    "spread": 1.35,
+    "weightKilo": 4.35
+  },
+  "reloads": {
+    "magFed": false,
+    "maxAmmo": 5,
+    "reloadAmount": 1,
+    "preReloadPauseTicks": 13,
+    "interReloadPauseTicks": 18,
+    "openBolt": true
+  },
+  "projectile": {
+    "item": "tac:bullet_308",
+    "visible": false,
+    "gravity": true,
+    "damage": 24,
+    "size": 0.2,
+    "speed": 110.0,
+    "life": 4
+  },
+  "sounds": {
+    "fire": "tac:item.fire_mosin-nagant",
+    "reload": "tac:item.pistol.reload",
+    "cock": "tac:item.pistol.cock",
+    "silencedFire": "tac:item.fire_mosin-nagant"
+  },
+  "display": {
+    "flash": {
+      "yOffset": 5.1,
+      "zOffset": -29.90,
+      "size": 1.3
+    }
+  },
+  "modules": {
+    "zoom": [{
+      "fovModifier": 0.65,
+      "yOffset": 7.20,
+      "zOffset": -6.4
+    }],
+    "attachments": {
+      "scope": {
+        "yOffset": 6.95,
+        "zOffset": 6.0,
+        "scale": 1.25
+      }
+    }
+  }
+}
