+++ src/main/resources/moonshine.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": true,
+  "package": "moonshine.core.mixin",
+  "compatibilityLevel": "JAVA_8",
+  "minVersion": "0.8",
+  "refmap": "moonshine.refmap.json",
+  "mixins": [],
+  "client": [],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
