--- src/main/resources/tac_extra.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "tac_extra.core.mixin",
-  "compatibilityLevel": "JAVA_8",
-  "minVersion": "0.8",
-  "refmap": "tac_extra.refmap.json",
-  "mixins": [],
-  "client": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
